How to use the Financial Calculator
To run this program, you need to download the Python program onto your computer operating system so that this Financial Calculator can be executed.
Link to download Python on your computer( Mac, Linux, Windows etc.) https://www.python.org/downloads/
Once Python is downloaded, you need IDLE(An Integrated Development Environment). IDLE is a safe environment to view and run your code. Python comes with built-in IDLE that can be accessed easily by opening the IDLE files from the program with the Python interpreter.
Once IDLE is opened it will display a "Shell" window, from the window, click "File", then click "Open" from the Toolbar(top of the Window) to open the Financial Calculator.
To run the Financial Calculator, click "Run" from the top Toolbar and check the output that shows in the Python shell Window.
You will be asked to input information to select a calculator